Move utils to separate files
